requirements.txt
# Install virtualenv if you don't have it
pip install virtualenv

# Create a virtual environment
python -m virtualenv kivy-env

# Activate the virtual environment
# On Windows:
kivy-env\Scripts\activate
# On macOS/Linux:
source kivy-env/bin/activate

pip install kivy